refactor(mcp-server): extract tool-registration/dispatch layer to its own module
Behavior-preserving split of src/mcp_server.py (1224 -> 837 lines). The
registration + dispatch-wrapping layer moves to a new src/tool_registration.py
so mcp_server.py is the server bootstrap and the hot path has a clear home.
Addresses the external review's "1,200-line server init" cognitive-load point.

Moved verbatim into src/tool_registration.py:
- get_tool_wrapper (the per-tool dispatch hot path: Wave 3a BEAM routing +
  Python fallback, metrics, TextContent unwrapping) and its wrapper cache
- auto_register_all_tools / _register_common_aliases
- _session_id_from_ctx (session-injection extractor)
- TOOLS_NEEDING_SESSION_INJECTION / EXTRA_ARGUMENT_PASSTHROUGH_TOOLS

Dependency injection breaks the circular import: the two registrars now take
the FastMCP instance as a parameter (auto_register_all_tools(mcp) /
_register_common_aliases(mcp)) instead of importing `mcp` back from mcp_server.
mcp_server builds `mcp`, then imports and calls them. Removed the now-unused
imports from mcp_server (dispatch_tool, wave3a routing/proxy, TOOL_CALLS_*,
Context, json, Dict).

Patch-location update: get_tool_wrapper's closure now binds tool_registration's
module globals, so the one test that monkeypatches dispatch/proxy and clears the
wrapper cache (test_routing_table_miss_passthrough) now targets
src.tool_registration instead of src.mcp_server. No production behavior change —
the wrapper resolves the same objects either way; only the patch target moves.
test_lite_wire_surface reproduces the alias logic independently and is
unaffected.

Gate: ruff clean; test-cache.sh -> 11053 passed / 21 skipped; coverage 81.9%.
Import smoke test confirms 26 tools register identically (20 tools + 6 aliases).
