+++ tests/integration/test_wave_3a_routing_table.py
@@ -225,20 +225,20 @@ async def test_routing_table_miss_passthrough(
     Exercises ``get_tool_wrapper`` directly so the assertion observes the
     wrapper's actual behaviour. Earlier draft of this test only consulted
     ``wave3a_routing.get_route`` — that test would have passed even if the
-    routing-table integration in ``mcp_server.get_tool_wrapper`` were
-    deleted. FIND-R4 council fold: tests must observe the wrapper, not the
-    routing-table state.
+    routing-table integration in ``tool_registration.get_tool_wrapper`` were
+    deleted. Tests must observe the wrapper, not the routing-table state.
     """
     from src import wave3a_beam_proxy, wave3a_routing
-    from src import mcp_server
+    from src import tool_registration
 
     assert wave3a_routing.get_route("_wave3a_test_miss_tool") is None
     assert wave3a_routing.route_count() == 0
 
     # Spy on dispatch_tool and on the BEAM proxy so we can confirm exactly
-    # which path fired. ``patch("src.mcp_server.dispatch_tool")`` works
-    # because the wrapper closes over the module-level binding (FIND-A3
-    # hoisted ``_wave3a_get_route`` / ``_wave3a_proxy_to_beam`` similarly).
+    # which path fired. ``patch("src.tool_registration.dispatch_tool")`` works
+    # because the wrapper closes over the module-level binding in
+    # tool_registration (FIND-A3 hoisted ``_wave3a_get_route`` /
+    # ``_wave3a_proxy_to_beam`` similarly).
     dispatch_calls: List[Tuple[str, Dict[str, Any]]] = []
     proxy_calls: List[str] = []
 
@@ -258,15 +258,15 @@ async def test_routing_table_miss_passthrough(
             "proxy_to_beam invoked on routing-table-miss path"
         )
 
-    monkeypatch.setattr(mcp_server, "dispatch_tool", fake_dispatch)
+    monkeypatch.setattr(tool_registration, "dispatch_tool", fake_dispatch)
     monkeypatch.setattr(
-        mcp_server, "_wave3a_proxy_to_beam", fake_proxy
+        tool_registration, "_wave3a_proxy_to_beam", fake_proxy
     )
 
     # Wrappers are cached; clear the cache so a fresh wrapper picks up the
     # patched dispatch binding via closure-on-module.
-    mcp_server._tool_wrappers_cache.clear()
-    wrapper = mcp_server.get_tool_wrapper("_wave3a_test_miss_tool")
+    tool_registration._tool_wrappers_cache.clear()
+    wrapper = tool_registration.get_tool_wrapper("_wave3a_test_miss_tool")
     result = await wrapper(arg="value")
 
     # Python dispatch fired exactly once; BEAM proxy did NOT fire; no
@@ -279,7 +279,7 @@ async def test_routing_table_miss_passthrough(
 
     # Cleanup: pop the cached wrapper so the patched dispatch binding does
     # not leak into other tests in this module.
-    mcp_server._tool_wrappers_cache.pop("_wave3a_test_miss_tool", None)
+    tool_registration._tool_wrappers_cache.pop("_wave3a_test_miss_tool", None)
     # Silence unused-import lint while keeping the import-sanity probe.
     _ = wave3a_beam_proxy
